Software-SUNAT/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
cloud-sql-python-connector[pg8000]
pg8000
firebase-admin
